refactor: extract runtime cli validators
